+++ snippy/content/all_content.py
@@ -1,0 +1,30 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+#
+#  Snippy - command, solution, reference and code snippet manager.
+#  Copyright 2017-2018 Heikki J. Laaksonen  <laaksonen.heikki.j@gmail.com>
+#
+#  This program is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU Affero General Public License as published
+#  by the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU Affero General Public License for more details.
+#
+#  You should have received a copy of the GNU Affero General Public License
+#  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""all_content: All content management."""
+
+from snippy.constants import Constants as Const
+from snippy.content.base import ContentTypeBase
+
+
+class AllContent(ContentTypeBase):
+    """Content field management."""
+
+    def __init__(self, storage, run_cli=True):
+        super(AllContent, self).__init__(storage, run_cli, Const.ALL_CATEGORIES)