+++ snippy/snip.py
@@ -24,6 +24,7 @@ import sys
 from snippy.cause import Cause
 from snippy.config.config import Config
 from snippy.config.source.cli import Cli
+from snippy.content.all_content import AllContent
 from snippy.content.reference import Reference
 from snippy.content.snippet import Snippet
 from snippy.content.solution import Solution
@@ -76,19 +77,11 @@ class Snippy(object):
         elif Config.is_category_reference:
             Reference(self.storage).run()
         elif Config.is_category_all and Config.is_operation_search:
-            self._run_cli_all()
+            AllContent(self.storage).run()
         else:
             Cause.push(Cause.HTTP_BAD_REQUEST, 'content category \'all\' is supported only with search operation')
 
         Cause.print_message()
-
-    def _run_cli_all(self):
-        """Run operation for all categories."""
-
-        collection = Snippet(self.storage, run_cli=False).run()
-        collection.migrate(Solution(self.storage, run_cli=False).run())
-        collection.migrate(Reference(self.storage, run_cli=False).run())
-        collection.dump_term(Config.use_ansi, Config.debug_logs, Config.search_filter)
 
     def _run_server(self):
         """Run API server."""