+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ Bugfixes
 
 * Fix solution creation from empty template.
 * Fix resource object validation in POST.
+* Fix search from all categories with --all option.
 * All commits in `0.9.0`_.
 
 0.8.0 (2018-21-07)